Support for 'extract'
Output can be to stdout or named file.
Output format supports JSON and YAML

JSON output required quite a few hacks (YAML worked smoothly out of the
box). The problem is that MarshalJSON on the interfaces (such as
Component) simply failed due to UnsupportedType error (with no clue as
to which type was the problem). I ended up with helper functions to work
around the problem, but hopefully done in such a way that updates to the
schema will require only minimal changes. Specifically,
'lib/certifications/certification.go' has a helper function to take
Certification interface and coerce it into a specific type; from that
point, JSON output worked as expected.